add duplicate nh remove logic to zebra
- add duplicate nh remove logic to zebra
- fix counter reverse issue

Discussed-in: Merge-Request 13089747, URL: https://code.alibaba-inc.com/SONiC/alibgp/codereview/13089747

+++ lib/nexthop_group.h
@@ -145,12 +145,12 @@ extern void nexthop_group_json_nexthop(json_object *j,
 				       const struct nexthop *nh);
 
 /* Return the number of nexthops in this nhg */
-extern uint8_t nexthop_group_nexthop_num(const struct nexthop_group *nhg);
-extern uint8_t
+extern uint32_t nexthop_group_nexthop_num(const struct nexthop_group *nhg);
+extern uint32_t
 nexthop_group_nexthop_num_no_recurse(const struct nexthop_group *nhg);
-extern uint8_t
+extern uint32_t
 nexthop_group_active_nexthop_num(const struct nexthop_group *nhg);
-extern uint8_t
+extern uint32_t
 nexthop_group_active_nexthop_num_no_recurse(const struct nexthop_group *nhg);
 
 #ifdef __cplusplus
